+++ doc/CHANGELOG.md
@@ -11,6 +11,7 @@
 - `[Message]` Converted message tests to playwright. ([#1954](https://github.com/infor-design/enterprise-wc/issues/1954))
 - `[WeekView]` Added more input handling for `startHour`, `endHour`, and `timelineInterval`. ([#2446](https://github.com/infor-design/enterprise-wc/issues/2446))
 - `[WeekView]` Converted week view tests to playwright. ([#1993](https://github.com/infor-design/enterprise-wc/issues/1993))
+- `[MonthView]` Converted week month tests to playwright. ([#1956](https://github.com/infor-design/enterprise-wc/issues/1956))
 
 ## 1.2.0